54 be move folder and chats in other folders

+++ Backend/api_v2/app/services/folder_service.py
@@ -3,23 +3,21 @@
 from datetime import datetime, timezone
 from typing import Optional, List
 
-from fastapi import Depends
-
 from bson import ObjectId
 from pymongo.asynchronous.database import AsyncDatabase
 
+from .chat_service import ChatService
+
 from ..core import errors
-from ..models.user import UserInDB
-from ..models.folder import (
+from ..models.user_models import UserInDB
+from ..models.folder_models import (
     FolderCreate, 
     FolderInDB,
     )
 
-from .chat_service import ChatService
-
 SESSIONS_COLLECTION = "chat_sessions"
 FOLDERS_COLLECTION = "folders"
-MAX_FOLDER_DEPTH = 3
+
 
 class FolderService:
     def __init__(self, db: AsyncDatabase):
@@ -38,12 +36,10 @@ class FolderService:
     
     async def list_folders(
             self, 
-            user: UserInDB,  
-            folder_id: Optional[str] = None,
+            user: UserInDB,
     ) -> List[FolderInDB]:
         try:
-            query = {"owner_id": user.id, 
-                    "parent_folder_id": folder_id}
+            query = {"owner_id": user.id}
             cursor = self.folders.find(query).sort("created_at", 1)
 
             entries: List[FolderInDB] = []
@@ -54,6 +50,36 @@ class FolderService:
             return entries
         except TypeError:
             raise ValueError(errors.UNKNOWN_ERROR_0)
+
+
+    async def get_by_id(
+            self, 
+            user: UserInDB,
+            folder_id: str
+    ) -> FolderInDB:
+        doc = await self.folders.find_one(
+            {"_id": ObjectId(folder_id), "owner_id": user.id}
+        )
+        if not doc:
+            raise ValueError(errors.FOLDER_1000_NOT_FOUND)
+        
+        doc["_id"] = str(doc["_id"])
+        return FolderInDB(**doc)
+    
+    
+    async def get_by_title(
+            self,
+            user: UserInDB,
+            folder_title: str,
+    ) -> FolderInDB:
+        doc = await self.folders.find_one(
+            {"title": folder_title, "owner_id": user.id}
+        )
+        if not doc:
+            return None
+        
+        doc["_id"] = str(doc["_id"])
+        return FolderInDB(**doc)
     
 
     async def create_folder(
@@ -61,48 +87,19 @@ class FolderService:
             user: UserInDB, 
             folder_in: FolderCreate
     ) -> FolderInDB:
-        # time
         now = datetime.now(timezone.utc)
 
-        # check parent + calculate depth
-        parent_depth = 0
-        if folder_in.parent_folder_id:
-            parent = await self.get_by_id(
-                    user=user,
-                    folder_id=folder_in.parent_folder_id
-            )
-            if not parent:
-                raise ValueError(errors.FOLDER_1000_NOT_FOUND)
-            parent_depth = parent.depth
-
-        depth = parent_depth + 1
-        if depth > MAX_FOLDER_DEPTH:
-            raise ValueError(errors.FOLDER_1002_MAX_DEPTH_EXCEEDED)
+        folder = await self.get_by_title(user, folder_in.title)
+        if folder:
+            raise ValueError(errors.FOLDER_1001_NAME_EXISTS)
         
         doc = {
             "owner_id": user.id,
             "title": folder_in.title,
-            "parent_folder_id": folder_in.parent_folder_id,
-            "depth": depth,
             "created_at": now,
         }
         result = await self.folders.insert_one(doc)
         doc["_id"] = str(result.inserted_id)
-
-        return FolderInDB(**doc)
-
-
-    async def get_by_id(
-            self, 
-            user: UserInDB,
-            folder_id: str
-    ) -> Optional[FolderInDB]:
-        doc = await self.folders.find_one(
-            {"_id": ObjectId(folder_id), "owner_id": user.id}
-        )
-        if not doc:
-            return None
-        doc["_id"] = str(doc["_id"])
         return FolderInDB(**doc)
     
 
@@ -112,34 +109,16 @@ class FolderService:
             folder_id: str,
             chat_service: ChatService
     ) -> None:
-        # Find all child-chats and delete each of them
-        query = {"folder_id": folder_id}
-        child_chats = self.chats.find(query)
+        # Ensure folder exists and belongs to the user
+        await self.get_by_id(user, folder_id)
+
+        # Deleta all chats inside this folder
+        child_chats = self.chats.find({"folder_id": folder_id})
         async for cc in child_chats:
             try:
-                await chat_service.delete_chat(
-                    user=user,
-                    chat_id=str(cc["_id"])
-                )
-            except ValueError:
-                continue   
-
-        # Find all child-folders and delete each of them
-        query = {"parent_folder_id": folder_id}
-        child_folders = self.folders.find(query)
-        async for cf in child_folders:
-            try:
-                await self.delete_folder(
-                    user=user,
-                    folder_id=str(cf["_id"]),
-                    chat_service=chat_service
-                )
+                await chat_service.delete_chat(user=user, chat_id=str(cc["_id"]))
             except ValueError:
                 continue
         
-        # In the end, delete the folder we are currently in
-        await self.folders.delete_one(
-            {"_id": ObjectId(folder_id),
-             "owner_id": user.id}
-        )
-        return
+        # Delete folder itself
+        await self.folders.delete_one({"_id": ObjectId(folder_id), "owner_id": user.id})
